Resolves a bunch of compile errors (from GCC 4.7) including:
* Missed making destructor virtual, resulting in in it not being called.
* Unsigned int, int comparisons.
* A couple of unused variables
* Ignored open GL error return codes. (printing them to logger error
now)

+++ src/core/ComponentFramework/PVector.cpp
@@ -51,7 +51,7 @@ void Core::PVector::Release( int id )
 
 void* Core::PVector::Get( int id )
 {
-    assert( id >= 0 && id < m_count );
+    assert( id >= 0 && id < (int)m_count );
     return &(((unsigned char*)m_data)[id*m_typesize]);
 }
